Incluye validaciones

+++ src/vista/GuiRegistroConfederacion.java
@@ -1,15 +1,9 @@
 package vista;
 
 import entidades.Confederacion;
-import java.awt.BorderLayout;
-import java.awt.Container;
-import java.awt.FlowLayout;
-import java.awt.Frame;
-import java.awt.GridLayout;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-import java.io.File;
-import java.io.IOException;
+import java.awt.*;
+import java.awt.event.*;
+import java.io.*;
 import java.nio.file.Files;
 import java.util.ArrayList;
 import javax.swing.*;
@@ -80,6 +74,7 @@ public class GuiRegistroConfederacion extends JDialog implements ActionListener 
         panel2.setBorder(BorderFactory.createTitledBorder("Registradas: "));
 
         this.txtRegistrados = new JTextArea(5, 30);
+        this.txtRegistrados.setEditable(false);
         panel2.add(this.txtRegistrados);
 
         this.datos.add(panel, BorderLayout.NORTH);
@@ -96,6 +91,7 @@ public class GuiRegistroConfederacion extends JDialog implements ActionListener 
         this.btnGuardar = new JButton("Guardar");
         this.btnGuardar.addActionListener(this);
         this.btnCancelar = new JButton("Cancelar");
+        this.btnCancelar.addActionListener(this);
 
         JPanel panel = new JPanel(new GridLayout(2, 1, 5, 5));
         panel.add(this.btnGuardar);
@@ -126,6 +122,7 @@ public class GuiRegistroConfederacion extends JDialog implements ActionListener 
     }
 
     public void cargarConfederaciones() {
+        this.txtRegistrados.setText(null);
         try {
             ArrayList<Confederacion> lista = this.modelo.leerConfederaciones();
             for (Confederacion c : lista) {
@@ -140,27 +137,46 @@ public class GuiRegistroConfederacion extends JDialog implements ActionListener 
     public void guardar() throws IOException {
 
         String nombre = this.txtNombre.getText();
+        if (nombre.isBlank()) {
+            throw new java.lang.NullPointerException("Se requiere el nombre de la Confederacion");
+        }
+
         String paises = this.txtPaises.getText();
+        if (paises.isBlank()) {
+            throw new java.lang.NullPointerException("Se requiere la descripcion de los paises");
+        }
+        
         int id = this.modelo.leerConfederaciones().size() + 1;
         Confederacion confe = new Confederacion(id, nombre, paises);
         this.modelo.registrarConfederacion(confe);
 
-        String name = this.banderaLeida.getName();
-        String ext = name.substring(name.indexOf("."), name.length());
-        File banderaDestino = new File("./src/imagenes/" + id +  ext);
+        if (this.banderaLeida != null && this.banderaLeida.exists() && this.banderaLeida.isFile()) {
+            String name = this.banderaLeida.getName();
+            String ext = name.substring(name.indexOf("."), name.length());
+            File banderaDestino = new File("./src/imagenes/" + id + ext);
 
-        Files.copy(this.banderaLeida.toPath(), banderaDestino.toPath());
+            Files.copy(this.banderaLeida.toPath(), banderaDestino.toPath());
+        }
 
     }
 
     public void leerLogo() {
         this.jf = new JFileChooser();
         int open = this.jf.showOpenDialog(null);
+        this.banderaLeida = null;
         if (open == JFileChooser.APPROVE_OPTION) {
             this.banderaLeida = this.jf.getSelectedFile();
             ImageIcon img = new ImageIcon(this.banderaLeida.getAbsolutePath());
             this.lbBandera.setIcon(img);
         }
+
+    }
+
+    public void limpiarCampos() {
+        this.txtNombre.setText(null);
+        this.txtPaises.setText(null);
+        this.cargarConfederaciones();
+        this.txtNombre.grabFocus();
 
     }
 
@@ -174,9 +190,16 @@ public class GuiRegistroConfederacion extends JDialog implements ActionListener 
             try {
                 this.guardar();
                 JOptionPane.showMessageDialog(this, "Registro exitoso", "Confirmacion", JOptionPane.INFORMATION_MESSAGE);
+                this.limpiarCampos();
             } catch (IOException ioe) {
-                System.out.println(ioe);
+                JOptionPane.showMessageDialog(this, ioe, "Error de archivo", JOptionPane.ERROR_MESSAGE);
+            }catch(NullPointerException np){
+                JOptionPane.showMessageDialog(this, np, "Error de validacion", JOptionPane.ERROR_MESSAGE);
             }
+        }
+
+        if (e.getSource() == this.btnCancelar) {
+            this.limpiarCampos();
         }
     }
